Reland: [soft navigations] Ignore non-attributed LCP before softnav detection
This relands a commit from pre-121 branch, with a number of fixes applied as
well as tests and changes to adapt to commits landed since that branch.

Original commit message:

This CL ignores LCPs that are not attributed to the soft navigation
task after a user interaction and before a soft navigation was
detected, as discussed in [1].

It also fixes a bug in soft navigation detection in the case where all
paints arrive before the URL was changed. Currently in such cases, a
soft navigation entry is not emitted. This CL fixes that.

[1] https://github.com/WICG/soft-navigations/issues/29#issuecomment-1831386977

Bug: 1505994
Change-Id: Ibf6d7d22375a1b0fc74e0fe31178831a738d8a46